double seems more precise

+++ src/mytimer/mytimer.c
@@ -12,7 +12,7 @@ void simpletimer_start_mql()
 {
 	simpletimer_cal_mql();
 }
-int simpletimer_end_mql() // ms
+double simpletimer_end_mql() // ms
 {
 	clock_t d=simpletimer_cal_mql();
 	return d*1.0/CLOCKS_PER_SEC*1000;
