+++ src/mytimer/mytimer.h
@@ -1,2 +1,2 @@
 void simpletimer_start_mql();
-int simpletimer_end_mql(); 
+double simpletimer_end_mql();
